[0.3.0] Order Processing Scripts Initial Release
The skeletons for publish_order which sends messages to the Pubsub topic and order_subscriber which listens to messages from the Pubsub subscription have now been implemented. The order message is currently hard coded into the publish_order file, but a Flask UI will be implemented next to manually enter order details.